refactor: 쿼리키 수정
